add app security, user is able to book/cancel tickets, create theatres and seats, get shows for theatre

+++ src/main/java/com/bookit/application/entity/Show.java
@@ -32,7 +32,7 @@ public class Show {
         return this.movie.getId();
     }
 
-    public Long getTheatreId() {
+    public Integer getTheatreId() {
         return this.theatre.getId();
     }
